+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ https://github.com/openoereb/oereb_client/milestone/5
 - Fix URL for static extract request
 - Fix loading indicator staying visible in map
 - Fix error on real estate without restriction
+- Respect specified layer opacity
 
 2.0.0-beta.2
 ************